// dashboard_react.client/src/types/CollectionResponse.ts
export interface CollectionResponse {
  id: string;
  name: string;
  description: string;
  nameView: string;
  isDeleted: boolean;
  isReadOnly: boolean;
  isVisible: boolean;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
  createdBy: string;
  updatedBy: string;
  deletedBy: string;
}